requirements.txt
asgiref==3.2.10
astroid==2.4.2
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1
django-cors-headers==3.4.0
django-ipware==3.0.1
django-seed==0.2.2
djangorestframework==3.11.1
Faker==4.1.1
gunicorn==20.0.4
idna==2.10
isort==4.3.21
itypes==1.2.0
jedi==0.17.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.1
openapi-codec==1.3.2
Pillow==7.2.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1
